requirements.txt
qiskit==1.4.2
randomname==0.2.1
httpx==0.27.0
dataclasses-json==0.6.4
dataclasses==0.6
pytimeparse==1.1.8
python-dotenv==1.0.1